size_t

+++ src/ds/test/map_test.cpp
@@ -140,7 +140,7 @@ TEST_CASE("persistent map operations")
 static const champ::Map<K, V, H> gen_map(size_t size)
 {
   champ::Map<K, V, H> map;
-  for (uint64_t i = 0; i < size; ++i)
+  for (size_t i = 0; i < size; ++i)
   {
     map = map.put(i, i);
   }
